[target-spec] reorganize code to prepare for errorless evaluation
Move the definitions and tests for `TargetSpec` and `Target` into
`parser.rs`.
